Module Tezos_benchmarks_proto_alpha.Global_constants_storage_benchmarks

val ns : Tezos_benchmark.Namespace.cons

val fv : string -> Tezos_benchmark.Free_variable.t

val assert_ok_lwt : ('a, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t -> 'a

val assert_ok : ('a, Tezos_base.TzPervasives.tztrace) Stdlib.result -> 'a

val seq_of_n_constants : int -> string -> (int, Tezos_protocol_alpha.Protocol.Michelson_v1_primitives.prim) Tezos_micheline.Micheline.node

seq_of_n_constants n hash generates a Seq filled with n constant primitives containing hash

val node_to_hash : ('a, Tezos_raw_protocol_alpha.Michelson_v1_primitives.prim) Tezos_micheline.Micheline.node -> string

Computes the b58check hash of a Micheline node as a string.

module Micheline_sampler : sig ... end

module Set_add : Benchmarks_proto.Benchmark.S

Cost model and benchmarks for set element addition from the OCaml stdlib.

module Set_elements : Benchmarks_proto.Benchmark.S

Cost model and benchmarks for set elements from the OCaml stdlib.

module Script_expr_hash_of_b58check_opt : Benchmarks_proto.Benchmark.S

Cost model and benchmarks for Script_expr_hash.of_b58_check_opt. Under the hood this function uses the Blake2b functor, which uses the HACL* crypto library.

module Global_constants_storage_expr_to_address_in_context : Benchmarks_proto.Benchmark.S

module Global_constants_storage_expand_models : sig ... end

Global_constants_storage.expand traverses a Micheline node, searching for constants and replacing them with their values retrieved from storage.