Module Tezos_benchmarks_proto_alpha.Interpreter_benchmarks

type ex_stack_and_kinstr =

| Ex_stack_and_kinstr : {

stack : 'a * 'b;

kinstr : ('a, 'b, 'c, 'd) Tezos_protocol_alpha.Protocol.Script_typed_ir.kinstr;

} -> ex_stack_and_kinstr

type ex_stack_and_continuation =

| Ex_stack_and_cont : {

stack : 'a * 'b;

cont : ('a, 'b, 'c, 'd) Tezos_protocol_alpha.Protocol.Script_typed_ir.continuation;

} -> ex_stack_and_continuation

type ex_value =

| Ex_value : {

value : 'a;

ty : 'a Tezos_protocol_alpha.Protocol.Script_typed_ir.ty;

} -> ex_value

val sf : ('a, unit, string) Stdlib.format -> 'a

val eos : Tezos_protocol_alpha.Protocol.Script_typed_ir.empty_cell * Tezos_protocol_alpha.Protocol.Script_typed_ir.empty_cell

val info_and_name : intercept:bool -> ?salt:string -> string -> string * string

module Default_boilerplate : sig ... end

module Default_config : sig ... end

val make_default_samplers : ?algo:[ `Algo of Tezos_crypto.Signature.algo | `Default ] -> Tezos_benchmark_alpha.Michelson_samplers_parameters.t -> (module Tezos_benchmark_alpha.Michelson_samplers.S)

val benchmark_from_kinstr_and_stack : ?amplification:int -> Tezos_protocol_alpha.Protocol.Alpha_context.context -> Tezos_protocol_alpha.Protocol.Script_interpreter.step_constants -> ex_stack_and_kinstr -> Interpreter_workload.ir_sized_step list Tezos_benchmark.Generator.benchmark

val make_benchmark : ?amplification:int -> ?intercept:bool -> ?salt:string -> name:Interpreter_workload.instruction_name -> kinstr_and_stack_sampler:(Default_config.config -> Stdlib.Random.State.t -> unit -> ex_stack_and_kinstr) -> unit -> Tezos_benchmark.Benchmark.t

val make_simple_benchmark : bef_top bef res_top res. ?amplification:int -> ?intercept:bool -> name:Interpreter_workload.instruction_name -> kinstr:('bef_top, 'bef, 'res_top, 'res) Tezos_protocol_alpha.Protocol.Script_typed_ir.kinstr -> unit -> Tezos_benchmark.Benchmark.t

val benchmark : ?amplification:int -> ?intercept:bool -> name:Interpreter_workload.instruction_name -> kinstr_and_stack_sampler:(Default_config.config -> Stdlib.Random.State.t -> unit -> ex_stack_and_kinstr) -> unit -> unit

val benchmark_with_stack_sampler : ?amplification:int -> ?intercept:bool -> name:Interpreter_workload.instruction_name -> kinstr:('a, 'b, 'c, 'd) Tezos_protocol_alpha.Protocol.Script_typed_ir.kinstr -> stack_sampler:(Default_config.config -> Stdlib.Random.State.t -> unit -> 'a * 'b) -> unit -> unit

val benchmark_with_fixed_stack : ?amplification:int -> ?intercept:bool -> name:Interpreter_workload.instruction_name -> stack:('a * 'b) -> kinstr:('a, 'b, 'c, 'd) Tezos_protocol_alpha.Protocol.Script_typed_ir.kinstr -> unit -> unit

val simple_benchmark_with_stack_sampler : ?amplification:int -> ?intercept_stack:('a * 'b) -> name:Interpreter_workload.instruction_name -> kinstr:('a, 'b, 'c, 'd) Tezos_protocol_alpha.Protocol.Script_typed_ir.kinstr -> stack_sampler:(Default_config.config -> Stdlib.Random.State.t -> unit -> 'a * 'b) -> unit -> unit

val simple_benchmark : ?amplification:int -> ?intercept_stack:('a * 'b) -> name:Interpreter_workload.instruction_name -> kinstr:('a, 'b, 'c, 'd) Tezos_protocol_alpha.Protocol.Script_typed_ir.kinstr -> unit -> unit

val benchmark_from_continuation : ?amplification:int -> Tezos_protocol_alpha.Protocol.Alpha_context.context -> Tezos_protocol_alpha.Protocol.Script_interpreter.step_constants -> ex_stack_and_continuation -> Interpreter_workload.ir_sized_step list Tezos_benchmark.Generator.benchmark

val make_continuation_benchmark : ?amplification:int -> ?intercept:bool -> ?salt:string -> name:Interpreter_workload.continuation_name -> cont_and_stack_sampler:(Default_config.config -> Stdlib.Random.State.t -> unit -> ex_stack_and_continuation) -> unit -> Tezos_benchmark.Benchmark.t

val continuation_benchmark : ?amplification:int -> ?intercept:bool -> ?salt:string -> name:Interpreter_workload.continuation_name -> cont_and_stack_sampler:(Default_config.config -> Stdlib.Random.State.t -> unit -> ex_stack_and_continuation) -> unit -> unit

val nat_of_positive_int : int -> Tezos_protocol_alpha.Protocol.Alpha_context.Script_int.n Tezos_protocol_alpha.Protocol.Alpha_context.Script_int.num

val adversarial_ints : Stdlib.Random.State.t -> Default_config.config -> int -> Tezos_protocol_alpha.Protocol.Script_int_repr.z Tezos_protocol_alpha.Protocol.Script_int_repr.num list

val raise_if_error : ('a, Tezos_error_monad.Error_monad.error Tezos_error_monad.Error_monad.TzTrace.trace) Stdlib.result -> 'a

Registration_section contains all interpreter benchmarks. The goal of a benchmark is to gather enough data to reliably estimate the parameters of the cost model associated to each instruction. In general, it can take several distinct benchmarks to properly cover all the execution paths.

In particular, for affine cost model, it is often worth estimating the intercept separately from the size-dependent coefficients.

module Registration_section : sig ... end