Module Samplers.Random_type

Samplers for random Michelson types.

val m_type : size:int -> ?blacklist:(Tezos_benchmark_alpha.Michelson_samplers.type_name -> bool) -> unit -> Tezos_protocol_alpha.Protocol.Script_typed_ir.ex_ty Tezos_benchmark.Base_samplers.sampler

m_type ~size ?blacklist samples a type containing exactly size constructors. The blacklist is a predicate which can be used to discard some unwanted cases.

val m_comparable_type : size:int -> Tezos_protocol_alpha.Protocol.Script_ir_translator.ex_comparable_ty Tezos_benchmark.Base_samplers.sampler

m_comparable_type ~size samples a comparable type containing exactly size constructors.